Remove RepositoryList
